CDS extractor index-files.sh compile to stdout
Attempt to resolve unit test failures for PR #170 of the
`advanced-security/codeql-sap-js` repo.

Updates the (soon to be replaced) shell-based version of the CDS
extractor as an attempted workaround for a change in cds compiler
behavior when the `-o` or `--dest` options are set for the `cds compile`
CLI command. Forces the cds compmiler to output to the desired .cds.json
file path via stdout.