Change sap.ui.require in the test cases to sap.ui.define
The importing function sap.ui.require mixes
CommonJS-style import and AMD module import,
making it hard to model. Besides, we do not directly
model it as of now; the previous calls to `sap.ui.require`
were captured only because the definition of
SapDefineModule was broken (it was too lax).

Therefore, change the tests to use sap.ui.define
instead of sap.ui.require; sap.ui.require used in
an AMD style is essentially a sap.ui.define without
defining a module in itself. Therefore, for the purposes
of the tests sap.ui.define in place of sap.ui.require
will still achieve what the tests intend to test.